+++ src/cheats_advanced_asset.rs
@@ -66,7 +66,7 @@ fn adv_asset_transfer_tx(
     });
     let args = rmp_serialize(&args).unwrap();
 
-    common::build_transaction(
+    common::build_unit_transaction(
         caller,
         network,
         asset_account,
@@ -102,7 +102,7 @@ fn adv_asset_init_tx(
     });
     let args = rmp_serialize(&args).unwrap();
 
-    common::build_transaction(
+    common::build_unit_transaction(
         caller,
         network,
         asset_account,
@@ -127,7 +127,7 @@ pub fn adv_asset_mint_tx(
     });
     let args = rmp_serialize(&args).unwrap();
 
-    common::build_transaction(
+    common::build_unit_transaction(
         caller,
         network,
         asset_account,
@@ -150,7 +150,7 @@ fn adv_asset_balance_tx(
         None => vec![],
     };
 
-    common::build_transaction(
+    common::build_unit_transaction(
         caller,
         network,
         asset_account,